feat: add switcher for mobile
