+++ src/components/Header/Header.tsx
@@ -1,15 +1,13 @@
 import { useEffect, useState } from 'react';
-import { useTranslation } from 'react-i18next';
-
 import { getToken } from '../../utils/functions/localStorage';
 import './Header.scss';
 import { BurgerMenu } from './BurgerMenu';
 import AuthLinks from './AuthLinks';
 import UnAuthLinks from './UnAuthLinks';
+import Switcher from '../Switcher';
 
 const Header = () => {
   const token = getToken();
-  const { i18n } = useTranslation();
   const [scrollPosition, setScrollPosition] = useState<number>(0);
 
   const handleScroll = () => {
@@ -25,10 +23,6 @@ const Header = () => {
     };
   }, []);
 
-  const handleLangChange = (e: React.ChangeEvent<HTMLSelectElement>) => {
-    i18n.changeLanguage(e.target.value);
-  };
-
   return (
     <header
       className="header"
@@ -41,19 +35,7 @@ const Header = () => {
           <nav className="header__nav">
             <ul className="header__nav_list">
               {!token ? <UnAuthLinks /> : <AuthLinks />}
-              <li
-                className="header__nav_item"
-                style={{ display: 'flex', flexDirection: 'row', alignItems: 'center' }}
-              >
-                <select
-                  className="header__lang"
-                  value={localStorage.getItem('i18nextLng') || 'en'}
-                  onChange={handleLangChange}
-                >
-                  <option value="en">EN</option>
-                  <option value="ru">RU</option>
-                </select>
-              </li>
+              <Switcher />
             </ul>
           </nav>
           <div className="burger__wrap">
